requirements.txt
audioop-lts; python_version>='3.13'
discord.py==2.4.0
python-dotenv==1.0.1
PyNaCl==1.5.0
requests==2.32.3
streamscrobbler @ git+https://github.com/CGillen/streamscrobbler-python.git@master
validators==0.34.0
psutil==5.9.8
fastapi==0.112.2
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
websockets==12.0
